Grunt concat, qunit, jshint, and uglify tasks
I created a package.json file and Gruntfile.js borrowing heavily from
the recommendations in the [sample][0]. I then loaded the following
dependencies:

* grunt: Runs tasks.
* grunt-contrib-concat: Concatenates source files into one file.
* grunt-contrib-qunit: Runs QUnit tests headlessly with PhantomJS.
* grunt-contrib-jshint: Validates syntax.
* grunt-contrib-uglify: Minifies source.

Along the way, I had to re-work the Sinon portion of the XHR tests. It
actually ended up simpler afterwards. I also had to add a bunch of
semicolons.

In addition to the respective default tasks for each of these
dependencies, there are two custom tasks, which will be part of the
development workflow:

* grunt test: Validate syntax and run tests headlessly.
* grunt: Validate syntax, run tests, concatenate, and minify. This is
  the default task.

0: http://gruntjs.com/sample-gruntfile