Object.prototype.$forEach()
Looping over the keys and values of an object requires that you make
sure each key is a property of the object and not a property from
somewhere on the prototype chain. Like so:

for (var key in object) {
  if (object.hasOwnProperty(key)) {
    // Do stuff...
  }
}

Now, you can simply call $forEach() to loop over the keys and values
of an object more idiomatically without checking hasOwnProperty().

object.$forEach(function(key, value) {
  // Do stuff...
})

Closes #23, https://github.com/canaryup/sprinkles/issues/23.